test/data/tree-construction/regression.dat

#data
<html><body><table><tr><td><div><b></div>X<table></table></td></tr></table><table></table><script type="text/javascript"></script></body></html>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <div>
|               <b>
|             <b>
|               "X"
|               <table>
|     <table>
|     <script>
|       type="text/javascript"

#data
<font><center></font>
#errors
#document
| <html>
|   <head>
|   <body>
|     <font>
|     <center>
|       <font>
